feat: programa principal criado para testar as funcionalidades da Agenda

+++ lab03/agenda.h
@@ -14,7 +14,7 @@ class Agenda {
 
     private:
         vector<string> nomes;
-        int tamanhoListaNomes;
+        long unsigned int tamanhoListaNomes;
 
 };
